requirements.txt
certifi==2020.12.5
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.10
Flask-MySQLdb==0.2.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql-connector-python==8.0.23
mysqlclient==2.0.3
protobuf==3.14.0
six==1.15.0
SQLAlchemy==1.3.22
Werkzeug==1.0.1
wincertstore==0.2
WTForms==2.3.3